faster path normalization after profiling
the use of charCodeAt was inspired by Node source. but of course that
runs on V8 and ExtendScript runs on SpiderMonkey. in the latter case,
detecting windows drive names is quicker with a regex.

it's extremely unlikely this is actually significant, but it at least
cuts down on code and dispels hunches